requirements.txt
torch
torchvision
opencv-python
matplotlib
PyMuPDF
Pillow